+++ components/cascader/Menu.js
@@ -39,64 +39,67 @@ const Menu = forwardRef(
         // currentDeep.current = deep
         if ((isFiltered && value.length > deep) || !isFiltered) {
           menus.push(
-            <ul
-              id="hi-cascader-menu"
-              className={`hi-cascader-menu theme__${theme} hi-cascader_transition`}
-              key={deep}
-              style={{ width: 'auto' }}
-            >
-              {_currentOptions.length === 0 && <Loading size="small" />}
-              {_currentOptions.map((option, index) => {
-                const optionValue = option[valueKey]
-                const hasChildren = Array.isArray(option[childrenKey])
-                const isExpanded = hasChildren && optionValue === currentValue
-                const expandIcon = 'icon-right'
-                const { _path } = option
-                const _deep = deep
-                const optionValues = option.jointOption ? optionValue : getOptionValues(value, optionValue, deep) // jointOption为true代表搜索拼接出来的option，直接取value即可
-                if (isExpanded) {
-                  currentOptions = option[childrenKey]
-                }
-                return (
-                  <li
-                    className={classNames('hi-cascader-menu__item', {
-                      'hi-cascader-menu__item-expanded': hasChildren,
-                      'hi-cascader-menu__item-disabled': !!option.disabled,
-                      'hi-cascader-menu__item-active': currentValue === optionValue,
-                      'hi-cascader-menu__item-focus': String(_path) === focusOptionIndex && targetByKeyDown.current,
-                      'hi-cascader-menu__item--isFiltered': isFiltered && !option.hightlight,
-                      'hi-cascader-menu__item--path': isFiltered && value.includes(option.id)
-                    })}
-                    onClick={(e) => {
-                      e.stopPropagation()
-                      if (!option.disabled) {
-                        onSelect(optionValues, hasChildren)
-                        targetByKeyDown.current = false
-                        setFocusOptionIndex(_path)
-                        currentDeep.current = _deep
-                      }
-                    }}
-                    onMouseEnter={(e) => {
-                      e.stopPropagation()
-                      if (!option.disabled && expandTrigger === 'hover') {
-                        onHover(optionValues, hasChildren)
-                        targetByKeyDown.current = false
-                        setFocusOptionIndex(_path)
-                        currentDeep.current = _deep
-                      }
-                    }}
-                    key={optionValue + option[labelKey]}
-                  >
-                    <span className={classNames(`hi-cascader-menu__item--label`)}>
-                      {isFiltered && option.hightlight
-                        ? (option.hightlight || []).map((item, index) => <span key={index}>{item}</span>)
-                        : option[labelKey]}
-                    </span>
-                    {hasChildren && <i className={classNames('hi-cascader-menu__item--icon', 'hi-icon', expandIcon)} />}
-                  </li>
-                )
-              })}
-            </ul>
+            <Loading visible={_currentOptions.length === 0} key={deep}>
+              <ul
+                id="hi-cascader-menu"
+                className={`hi-cascader-menu theme__${theme} hi-cascader_transition`}
+                key={deep}
+                style={{ width: 'auto' }}
+              >
+                {_currentOptions.map((option, index) => {
+                  const optionValue = option[valueKey]
+                  const hasChildren = Array.isArray(option[childrenKey])
+                  const isExpanded = hasChildren && optionValue === currentValue
+                  const expandIcon = 'icon-right'
+                  const { _path } = option
+                  const _deep = deep
+                  const optionValues = option.jointOption ? optionValue : getOptionValues(value, optionValue, deep) // jointOption为true代表搜索拼接出来的option，直接取value即可
+                  if (isExpanded) {
+                    currentOptions = option[childrenKey]
+                  }
+                  return (
+                    <li
+                      className={classNames('hi-cascader-menu__item', {
+                        'hi-cascader-menu__item-expanded': hasChildren,
+                        'hi-cascader-menu__item-disabled': !!option.disabled,
+                        'hi-cascader-menu__item-active': currentValue === optionValue,
+                        'hi-cascader-menu__item-focus': String(_path) === focusOptionIndex && targetByKeyDown.current,
+                        'hi-cascader-menu__item--isFiltered': isFiltered && !option.hightlight,
+                        'hi-cascader-menu__item--path': isFiltered && value.includes(option.id)
+                      })}
+                      onClick={(e) => {
+                        e.stopPropagation()
+                        if (!option.disabled) {
+                          onSelect(optionValues, hasChildren)
+                          targetByKeyDown.current = false
+                          setFocusOptionIndex(_path)
+                          currentDeep.current = _deep
+                        }
+                      }}
+                      onMouseEnter={(e) => {
+                        e.stopPropagation()
+                        if (!option.disabled && expandTrigger === 'hover') {
+                          onHover(optionValues, hasChildren)
+                          targetByKeyDown.current = false
+                          setFocusOptionIndex(_path)
+                          currentDeep.current = _deep
+                        }
+                      }}
+                      key={optionValue + option[labelKey]}
+                    >
+                      <span className={classNames(`hi-cascader-menu__item--label`)}>
+                        {isFiltered && option.hightlight
+                          ? (option.hightlight || []).map((item, index) => <span key={index}>{item}</span>)
+                          : option[labelKey]}
+                      </span>
+                      {hasChildren && (
+                        <i className={classNames('hi-cascader-menu__item--icon', 'hi-icon', expandIcon)} />
+                      )}
+                    </li>
+                  )
+                })}
+              </ul>
+            </Loading>
           )
           deep++
         }
