+++ docs/demo/cascader/section-advanced.jsx
@@ -82,7 +82,8 @@ class Demo extends React.Component {
   }
 }`,
     opt: ['自定义字段名']
-  }, {
+  },
+  {
     code: `import React from 'react'
 import Cascader from '@hi-ui/hiui/es/cascader'\n
 class Demo extends React.Component {
@@ -157,109 +158,15 @@ class Demo extends React.Component {
   }
 }`,
     opt: ['自定义显示']
-  }, {
+  },
+  {
     code: `import React from 'react'
 import Cascader from '@hi-ui/hiui/es/cascader'\n
 class Demo extends React.Component {
   constructor () {
     super()
     this.state = {
-      options: [
-        {
-          id: '手机',
-          content: '手机',
-          children: [
-            {
-              id: '小米',
-              content: '小米',
-              children: [
-                {
-                  id: '小米1',
-                  content: '小米1'
-                },
-                {
-                  id: '小米2',
-                  content: '小米2',
-                  disabled: true
-                },
-                {
-                  id: '小米3',
-                  content: '小米3'
-                },
-                {
-                  id: '小米4',
-                  content: '小米4'
-                },
-                {
-                  id: '小米5',
-                  content: '小米5'
-                },
-                {
-                  id: '小米6',
-                  content: '小米6'
-                },
-                {
-                  id: '小米7',
-                  content: '小米7'
-                },
-                {
-                  id: '小米8',
-                  content: '小米8'
-                }
-              ]
-            },
-            {
-              id: '红米',
-              content: '红米',
-              children: [
-                {
-                  id: '红米1',
-                  content: '红米1'
-                },
-                {
-                  id: '红米2',
-                  content: '红米2'
-                },
-                {
-                  id: '红米3',
-                  content: '红米3'
-                },
-                {
-                  id: '红米4',
-                  content: '红米4'
-                }
-              ]
-            }
-          ]
-        },
-        {
-          id: '电视',
-          content: '电视',
-          children: [
-            {
-              id: '小米电视4A',
-              content: '小米电视4A'
-            },
-            {
-              id: '小米电视4C',
-              content: '小米电视4C'
-            },
-            {
-              id: '小米电视4X',
-              content: '小米电视4X'
-            },
-            {
-              id: '小米电视4',
-              content: '小米电视4'
-            }
-          ]
-        },
-        {
-          id: 'mix',
-          content: 'Mix',
-          children: []
-        }
-      ]
+      options: []
     }
   }
   render(){
@@ -270,10 +177,41 @@ class Demo extends React.Component {
         }}
         data={this.state.options}
         style={{ width: 240 }}
+        onOpen={()=>{
+          const {options} = this.state
+          if(options.length === 0){
+            setTimeout(()=>{
+              this.setState({
+                options: [
+                  {
+                    id: '手机',
+                    content: '手机',
+                    children: [
+                      {
+                        id: '手机2',
+                        content: '手机2',
+                      }
+                    ]
+                  },
+                  {
+                    id: '平板',
+                    content: '平板'
+                  },
+                  {
+                    id: 'mix',
+                    content: 'Mix',
+                    children: []
+                  }
+                ]
+              })
+            }, 2000)
+          }
+        }}
         onActiveItemChange={values=>{
           if(values[0] == 'mix') {
             setTimeout(()=>{
-              this.state.options[2].children = [
+              const _options = this.state.options
+              _options[2].children = [
                 {
                   id: 'mix1',
                   content: 'Mix1'
@@ -285,7 +223,9 @@ class Demo extends React.Component {
                   content: 'Mix3'
                 }
               ]
-              this.forceUpdate()
+              this.setState({
+                options: _options
+              })
             }, 1000)
           }
         }}
@@ -294,7 +234,8 @@ class Demo extends React.Component {
   }
 }`,
     opt: ['动态加载选项']
-  }, {
+  },
+  {
     code: `import React from 'react'
 import Cascader from '@hi-ui/hiui/es/cascader'\n
 class Demo extends React.Component {
@@ -446,12 +387,6 @@ class Demo extends React.Component {
 ]
 
 const DemoBasic = () => (
-  <DocViewer
-    code={code}
-    scope={{ Cascader }}
-    prefix={prefix}
-    desc={desc}
-    rightOptions={rightOptions}
-  />
+  <DocViewer code={code} scope={{ Cascader }} prefix={prefix} desc={desc} rightOptions={rightOptions} />
 )
 export default DemoBasic
